+++ src/main/java/com/jackdaw/chatwithnpc/npc/VillagerNPCEntity.java
@@ -1,7 +1,5 @@
 package com.jackdaw.chatwithnpc.npc;
 
-import com.jackdaw.chatwithnpc.prompt.Builder;
-import com.jackdaw.chatwithnpc.prompt.Prompt;
 import net.minecraft.entity.passive.VillagerEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.Text;
