+++ src/main/java/com/jackdaw/chatwithnpc/prompt/Builder.java
@@ -3,28 +3,27 @@ package com.jackdaw.chatwithnpc.prompt;
 import com.jackdaw.chatwithnpc.auxiliary.configuration.SettingManager;
 import com.jackdaw.chatwithnpc.group.Group;
 import com.jackdaw.chatwithnpc.group.GroupEvent;
+import com.jackdaw.chatwithnpc.group.GroupManager;
+import com.jackdaw.chatwithnpc.npc.NPCEntityManager;
 import com.jackdaw.chatwithnpc.npc.NPCEntity;
 import com.jackdaw.chatwithnpc.npc.Record;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 
 public class Builder {
     String npcName = "NPC";
-
+    String npcGroup = "Global";
     String npcType = "Minecraft:villager";
     String npcCareer = "unemployed";
     String npcBasicPrompt = "I'm an NPC.";
-    ArrayList<Group> groups = new ArrayList<>();
-    Record npcMessageRecord = new Record();
 
-    public Builder setNpc(@NotNull NPCEntity npc) {
+    public Builder setNpc(String npcName) {
+        NPCEntity npc = NPCEntityManager.getNPCEntity(npcName);
         this.npcName = npc.getName();
+        this.npcGroup = npc.getGroup();
         this.npcType = npc.getType();
         this.npcCareer = npc.getCareer();
         this.npcBasicPrompt = npc.getBasicPrompt();
-        this.groups = npc.getGroup().getParentGroups();
-        this.npcMessageRecord = npc.getMessageRecord();
         return this;
     }
 
@@ -43,16 +42,6 @@ public class Builder {
         return this;
     }
 
-    public Builder setGroups(ArrayList<Group> groups) {
-        this.groups = groups;
-        return this;
-    }
-
-    public Builder setNpcMessageRecord(Record npcMessageRecord) {
-        this.npcMessageRecord = npcMessageRecord;
-        return this;
-    }
-
     public Prompt build() {
         String systemMessage = buildSystemMessage();
         String historyMessage = buildHistoryMessage();
@@ -63,7 +52,7 @@ public class Builder {
         String npcbasicPrompt = "You are an NPC with type `" + npcType + "` and named `" + npcName + "`. ";
         String groupInitialPrompt = "You are living in places as listed: ";
         ArrayList<String> groupPrompt = new ArrayList<>();
-        for (Group group : groups) {
+        for (Group group : GroupManager.getParentGroups(npcGroup)) {
             // connect the group.permanentPrompt to the groupPrompt with ",";
             StringBuilder prompt = new StringBuilder();
             if (group.getName().equals("Global")) {
@@ -91,7 +80,7 @@ public class Builder {
     String buildHistoryMessage() {
         StringBuilder messageBuilder = new StringBuilder();
         messageBuilder.append("This is the history of our conversation: ");
-        for (Record.Message message : npcMessageRecord.getTreeMap().values()) {
+        for (Record.Message message : NPCEntityManager.getNPCEntity(npcName).getMessageRecord().getTreeMap().values()) {
             if (message.role == Record.Role.NPC) {
                 messageBuilder.append("You said: " + message.getMessage() + ". ");
             } else {
